Resolve issues with postUrl with list; add tests
postUrl assumed that the response would always be a HashMap.
This PR adds the ability return a HashMap when an ArrayList is received.

It also includes modifications to the containerZone functions to support
the created HashMap from ArrayList.

Adds additional mocks for Jenkins/API endpoints. Also a script that will automate the
running of the tests.